cmd: split supervise_test.go for Darwin portability (review-code blocker)
review-code on round 1 flagged that cmd/supervise_test.go used
Linux-only syscall constants (TIOCSPTLCK, TIOCGPTN) and the
Setctty/Ctty fields of syscall.SysProcAttr unconditionally, which
broke 'go test ./cmd/' on Darwin even though Linux-only CI did not
catch it. The cmd package is built and tested on Darwin too (see the
existing _darwin.go siblings), so this was a real portability
regression.

Three-file split:

- cmd/supervise_test.go (portable, all GOOS): early-return tests
  (TestSuperviseChild_NilCmdReturnsImmediately,
  TestSuperviseChild_UnstartedCmdReturnsImmediately) plus the
  superviseHelperEnvVar constant. The constant must be reachable on
  all GOOS because the package-wide TestMain in killsidecar_test.go
  dispatches on it unconditionally.

- cmd/supervise_linux_test.go (//go:build linux): everything that
  needs Linux-specific syscalls or SysProcAttr fields — all six
  signal-forwarding and foreground-pgid tests, the helper-subprocess
  machinery (runSuperviseHelper, runSuperviseSignalHelper,
  runSuperviseForegroundHelper, startSuperviseHelper*,
  signalHelperAndAwaitExit, openPTYPair, boolFlag).

- cmd/supervise_other_test.go (//go:build !linux): non-Linux stub
  for runSuperviseHelper that exits with sentinel code 99 if the
  helper is ever accidentally invoked on a non-Linux GOOS. Linux
  returns codes in the 0-28 range so 99 cannot collide with a real
  verdict.

The runtime.GOOS guards inside each Linux-only test were removed —
the build tag now does the same job declaratively and the tests no
longer need to skip themselves.

Verification:
- go build ./... (Linux) passes.
- GOOS=darwin go build ./... passes.
- GOOS=darwin go test -c -o /dev/null ./cmd/ passes (test binary
  links cleanly on Darwin).
- go test -race -run 'TestSuperviseChild' ./cmd/ — all 8 tests pass
  on Linux.
- go test ./... — full prism suite passes on Linux.
- nix build .#prism — passes.
- nix-build-prism-checked equivalent — passes (homeless-shelter
  gate satisfied).
