+++ modules/programs/prism/prism/cmd/supervise_test.go
@@ -1,6 +1,6 @@
 package cmd
 
-// Unit tests for supervise.go (SuperviseChild).
+// Portable (all-GOOS) unit tests for supervise.go (SuperviseChild).
 //
 // SuperviseChild owns three concerns for both the bwrap (Linux) and
 // sandbox-exec (Darwin) agent-run dispatch paths:
@@ -13,362 +13,46 @@ package cmd
 //  3. Early-return when the supplied *exec.Cmd is nil or has not been
 //     started (cmd.Process == nil).
 //
-// These tests target each of those concerns.
+// This file holds only the platform-agnostic early-return tests
+// (concern #3). The Linux-only tests for concerns #1 and #2 — which
+// rely on /dev/ptmx + TIOCSPTLCK/TIOCGPTN ioctls and the
+// Setctty/Ctty fields of syscall.SysProcAttr that exist only on
+// Linux — live in supervise_linux_test.go behind a //go:build linux
+// tag.
 //
-// Why the signal-forwarding tests re-exec the test binary as a helper
-// subprocess: SuperviseChild needs to receive a real OS signal (SIGTERM
-// / SIGINT / SIGHUP) at agent-run's PID and forward it to the child's
-// process group. We cannot raise those signals at the `go test` process
-// itself — the Go testing framework treats SIGTERM/SIGINT as termination
-// signals regardless of whether the test code has installed a
-// signal.Notify subscriber (the runtime's `_SigKill` flag for those
-// signals is asymmetric with the `signal` package's accept-all
-// semantics, so the test binary exits non-zero even when a Notify is
-// up). Instead, we launch a helper subprocess (this same binary, with
-// `PRISM_TEST_SUPERVISE_HELPER=1`) whose main runs SuperviseChild
-// against its own sleep child. The test then signals the helper's PID
-// and observes the helper's exit code — that proves the signal was
-// forwarded into the child pgid (which then propagated up via Wait).
-//
-// The Linux-only gating reflects the fact that the helper relies on
-// Setpgid + Kill(-pgid, sig) semantics, and on TIOCGPGRP/TIOCSPGRP
-// against a /dev/ptmx PTY for the foreground-pgid test.
+// The superviseHelperEnvVar constant is also declared here (rather
+// than in the Linux-only file) because the package-wide TestMain
+// dispatcher in killsidecar_test.go references it unconditionally.
+// The runSuperviseHelper function it dispatches to has a platform
+// split: the real implementation lives in supervise_linux_test.go,
+// and a stub returning a "not supported on this GOOS" sentinel lives
+// in supervise_other_test.go.
 
 import (
-	"bytes"
-	"fmt"
-	"io"
 	"os"
 	"os/exec"
-	"os/signal"
-	"runtime"
-	"sync/atomic"
-	"syscall"
 	"testing"
 	"time"
-	"unsafe"
 )
 
 // superviseHelperEnvVar gates the helper-subprocess entry point inside
-// the package-wide TestMain (see killsidecar_test.go). When set to "1",
-// the test binary skips the usual go-test entry and instead executes
-// the SuperviseChild helper described in runSuperviseHelper.
+// the package-wide TestMain (see killsidecar_test.go). When set to
+// "1", the test binary skips the usual go-test entry and instead
+// executes the SuperviseChild helper described in runSuperviseHelper.
+//
+// The constant is declared in this portable file so the TestMain
+// dispatcher in killsidecar_test.go can reference it on all GOOS; the
+// helper body itself is only meaningful on Linux (see
+// supervise_linux_test.go).
 const superviseHelperEnvVar = "PRISM_TEST_SUPERVISE_HELPER"
 
-// runSuperviseHelper implements the helper-subprocess body invoked by
-// the package-wide TestMain when superviseHelperEnvVar is set.
-//
-// Mode selection: PRISM_TEST_SUPERVISE_MODE selects between two
-// helper bodies:
-//
-//   - "" or "signal": signal-forwarding harness. Spawns `sleep 30`,
-//     runs SuperviseChild, reports its own PID via stdout, then blocks
-//     until a forwarded signal terminates the sleep child. Exit 0
-//     means the supervisor forwarded the signal.
-//
-//   - "foreground": foreground-pgid harness. Calls Setsid to become
-//     its own session leader, opens a /dev/ptmx PTY pair and acquires
-//     the slave as controlling terminal, captures the before-pgid,
-//     spawns a short-lived `sleep 0.3`, runs SuperviseChild against
-//     the slave PTY fd, then captures the after-pgid. Reports
-//     before/after via stdout as "before=%d after=%d\n". The parent
-//     test asserts before == after (restore happened).
-func runSuperviseHelper() int {
-	switch os.Getenv("PRISM_TEST_SUPERVISE_MODE") {
-	case "foreground":
-		return runSuperviseForegroundHelper()
-	default:
-		return runSuperviseSignalHelper()
-	}
-}
+// ── early-return paths (portable, all GOOS) ──────────────────────────────────
 
-// runSuperviseSignalHelper is the signal-forwarding helper body. See
-// the runSuperviseHelper doc-comment for the contract.
-func runSuperviseSignalHelper() int {
-	forwardWinch := os.Getenv("PRISM_TEST_SUPERVISE_FORWARD_WINCH") == "1"
-
-	cmd := exec.Command("sleep", "30")
-	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-	if err := cmd.Start(); err != nil {
-		fmt.Fprintf(os.Stderr, "helper: start sleep: %v\n", err)
-		return 10
-	}
-
-	// Tell the parent our PID so it can signal us.
-	fmt.Printf("pid=%d\n", os.Getpid())
-
-	// Use a non-TTY fd: stdin under exec.Cmd default is /dev/null,
-	// which is fine — tcsetpgrpForeground silently no-ops on non-TTY.
-	waitErr := SuperviseChild(cmd, int(os.Stdin.Fd()), SuperviseOpts{
-		ForwardWinch: forwardWinch,
-	})
-
-	// Expectation: a forwarded signal terminated the sleep child, so
-	// Wait returns a non-nil ExitError. Exit 0 means "test apparatus
-	// observed a forwarded signal". Exit 11 means "Wait returned nil"
-	// (child exited cleanly — unexpected). Exit 12 means "some other
-	// wait error" (also unexpected).
-	if waitErr == nil {
-		return 11
-	}
-	if _, ok := waitErr.(*exec.ExitError); !ok {
-		fmt.Fprintf(os.Stderr, "helper: unexpected wait error: %v\n", waitErr)
-		return 12
-	}
-	return 0
-}
-
-// runSuperviseForegroundHelper is the foreground-pgid helper body. See
-// the runSuperviseHelper doc-comment for the contract.
-//
-// Being a session leader is required because TIOCSPGRP fails with
-// EPERM if the calling process is not in the same session as the
-// terminal it is trying to control. The parent test sets up the
-// helper as follows:
-//
-//   - Opens /dev/ptmx in the parent, unlocks the slave, opens the
-//     slave path.
-//   - Launches the helper with Setsid: true + Setctty: true + Ctty: 0
-//     and the slave fd dup'd onto the helper's stdin (fd 0). The
-//     kernel then atomically: makes the helper a session leader,
-//     assigns the slave as the helper's controlling terminal, and
-//     starts the helper.
-//
-// This combination is exactly the shape that production uses for
-// agent-run (which inherits a controlling tty from its tmux pane).
-// Inside the helper we use fd 0 directly as the supervisor's stdin.
-func runSuperviseForegroundHelper() int {
-	// Confirm we are a session leader and our fd 0 is a TTY (both set
-	// by the parent's SysProcAttr). If not, the test apparatus is
-	// wrong.
-	sid, _, errno := syscall.Syscall(syscall.SYS_GETSID, 0, 0, 0)
-	if errno != 0 || int(sid) != os.Getpid() {
-		fmt.Fprintf(os.Stderr, "helper: not session leader: sid=%d pid=%d errno=%v\n", int(sid), os.Getpid(), errno)
-		return 20
-	}
-	slaveFd := 0
-
-	// Capture the before-pgid. As session leader with this PTY as our
-	// ctty, the foreground pgid should be our own pgid (== our pid).
-	var beforePgid int32
-	if _, _, errno := syscall.Syscall(syscall.SYS_IOCTL, uintptr(slaveFd), syscall.TIOCGPGRP, uintptr(unsafe.Pointer(&beforePgid))); errno != 0 {
-		fmt.Fprintf(os.Stderr, "helper: TIOCGPGRP (before) on fd %d: %v\n", slaveFd, errno)
-		return 25
-	}
-
-	// Spawn a short-lived child in its own process group.
-	cmd := exec.Command("sleep", "0.3")
-	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-	if err := cmd.Start(); err != nil {
-		fmt.Fprintf(os.Stderr, "helper: start sleep: %v\n", err)
-		return 26
-	}
-
-	// Run SuperviseChild against the slave PTY fd. It will:
-	//   1. tcsetpgrpForeground(slaveFd, cmd.Process.Pid)
-	//   2. cmd.Wait() — returns when `sleep 0.3` exits cleanly
-	//   3. tcsetpgrpRestore(slaveFd, beforePgid)
-	waitErr := SuperviseChild(cmd, slaveFd, SuperviseOpts{ForwardWinch: false})
-	if waitErr != nil {
-		fmt.Fprintf(os.Stderr, "helper: SuperviseChild returned: %v\n", waitErr)
-		return 27
-	}
-
-	// Capture the after-pgid. If SuperviseChild restored correctly,
-	// this should equal beforePgid.
-	var afterPgid int32
-	if _, _, errno := syscall.Syscall(syscall.SYS_IOCTL, uintptr(slaveFd), syscall.TIOCGPGRP, uintptr(unsafe.Pointer(&afterPgid))); errno != 0 {
-		fmt.Fprintf(os.Stderr, "helper: TIOCGPGRP (after): %v\n", errno)
-		return 28
-	}
-
-	// Diagnostic: explicitly call tcsetpgrpRestore now and report
-	// whether it succeeds. SuperviseChild already did this and
-	// discarded the error; if our explicit call also fails we know
-	// the kernel is rejecting TIOCSPGRP at this point in time.
-	restoreErr := tcsetpgrpRestore(slaveFd, int(beforePgid))
-	restoreErrStr := "<nil>"
-	if restoreErr != nil {
-		restoreErrStr = restoreErr.Error()
-	}
-
-	// Capture pgid AGAIN after the explicit restore.
-	var finalPgid int32
-	_, _, _ = syscall.Syscall(syscall.SYS_IOCTL, uintptr(slaveFd), syscall.TIOCGPGRP, uintptr(unsafe.Pointer(&finalPgid)))
-
-	fmt.Printf("before=%d after=%d final=%d child_pid=%d restore_err=%q\n", beforePgid, afterPgid, finalPgid, cmd.Process.Pid, restoreErrStr)
-
-	// Detach from the controlling terminal before returning so the
-	// kernel does not send SIGHUP to us when our session leader (us)
-	// exits with a still-attached ctty. Without this the helper can
-	// catch SIGHUP between the final Printf and os.Exit, causing the
-	// parent to see exit-signal=SIGHUP instead of exit=0.
-	if _, _, errno := syscall.Syscall(syscall.SYS_IOCTL, uintptr(slaveFd), syscall.TIOCNOTTY, 0); errno != 0 {
-		fmt.Fprintf(os.Stderr, "helper: TIOCNOTTY: %v (continuing)\n", errno)
-	}
-	return 0
-}
-
-// startSuperviseHelper launches the test binary as a signal-forwarding
-// helper subprocess and returns the *exec.Cmd plus the helper's own
-// PID (read from its stdout). The caller is responsible for sending
-// the signal under test and then waiting for the helper to exit.
-//
-// The helper is started in its own process group so the parent can
-// SIGKILL it via Kill(-pgid, SIGKILL) as a hard cleanup if the test
-// times out — without that the helper's `sleep 30` child would
-// otherwise outlive the test by 30 seconds.
-func startSuperviseHelper(t *testing.T, forwardWinch bool) (*exec.Cmd, int) {
-	t.Helper()
-
-	helper, stdout, _ := startSuperviseHelperWithMode(t, "signal", []string{
-		fmt.Sprintf("PRISM_TEST_SUPERVISE_FORWARD_WINCH=%s", boolFlag(forwardWinch)),
-	})
-
-	// Read the first line ("pid=%d\n") to get the helper's PID. This
-	// also synchronises with helper startup: we know the helper has
-	// reached the point in runSuperviseHelper just before
-	// SuperviseChild, so its signal.Notify subscription is about to
-	// come up. We add a tiny delay after the read to cover the small
-	// window between fmt.Printf and signal.Notify inside SuperviseChild.
-	var buf [64]byte
-	n, err := stdout.Read(buf[:])
-	if err != nil {
-		t.Fatalf("read helper pid line: %v", err)
-	}
-	var pid int
-	if _, err := fmt.Sscanf(string(buf[:n]), "pid=%d", &pid); err != nil {
-		t.Fatalf("parse helper pid line %q: %v", string(buf[:n]), err)
-	}
-
-	// Give the helper a moment to enter signal.Notify inside
-	// SuperviseChild. The pid line is printed *before* SuperviseChild
-	// is called, so a brief sleep here covers the gap.
-	time.Sleep(100 * time.Millisecond)
-
-	return helper, pid
-}
-
-// startSuperviseHelperWithMode launches the test binary as a helper
-// subprocess with the given PRISM_TEST_SUPERVISE_MODE value plus any
-// additional env entries. Returns the cmd, its stdout pipe, and a
-// pointer to a stderr buffer that the cleanup hook flushes to t.Logf
-// after the helper exits.
-//
-// For mode == "foreground" the parent opens a PTY pair and passes the
-// slave as the helper's stdin with Setsid + Setctty so the kernel
-// atomically makes the helper a session leader with the slave as its
-// controlling terminal. For other modes the helper just gets its own
-// process group via Setpgid (signal-forwarding mode does not need a
-// PTY).
-func startSuperviseHelperWithMode(t *testing.T, mode string, extraEnv []string) (*exec.Cmd, io.ReadCloser, *bytes.Buffer) {
-	t.Helper()
-
-	self, err := os.Executable()
-	if err != nil {
-		t.Fatalf("os.Executable: %v", err)
-	}
-
-	helper := exec.Command(self)
-	env := append([]string{}, os.Environ()...)
-	env = append(env,
-		superviseHelperEnvVar+"=1",
-		"PRISM_TEST_SUPERVISE_MODE="+mode,
-	)
-	env = append(env, extraEnv...)
-	helper.Env = env
-
-	if mode == "foreground" {
-		// Open a PTY pair in the parent and pass the slave as the
-		// helper's stdin. Combined with Setsid+Setctty in SysProcAttr
-		// this gives the helper a controlling terminal that
-		// SuperviseChild can drive with TIOCSPGRP.
-		master, slave, err := openPTYPair()
-		if err != nil {
-			t.Skipf("open PTY pair: %v (no PTY available in this environment)", err)
-		}
-		helper.Stdin = slave
-		helper.SysProcAttr = &syscall.SysProcAttr{
-			Setsid:  true,
-			Setctty: true,
-			Ctty:    0, // index into helper.Files: stdin = slave
-		}
-		t.Cleanup(func() {
-			master.Close()
-			slave.Close()
-		})
-	} else {
-		helper.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-	}
-
-	stdout, err := helper.StdoutPipe()
-	if err != nil {
-		t.Fatalf("StdoutPipe: %v", err)
-	}
-	stderrBuf := &bytes.Buffer{}
-	helper.Stderr = stderrBuf
-
-	if err := helper.Start(); err != nil {
-		t.Fatalf("start helper: %v", err)
-	}
-	t.Cleanup(func() {
-		if helper.Process != nil {
-			// For Setsid helpers the helper IS its own session and
-			// pgid leader, so -pid targets the whole helper group.
-			_ = syscall.Kill(-helper.Process.Pid, syscall.SIGKILL)
-		}
-		if stderrBuf.Len() > 0 {
-			t.Logf("helper stderr: %s", stderrBuf.String())
-		}
-	})
-
-	return helper, stdout, stderrBuf
-}
-
-// openPTYPair opens /dev/ptmx and the corresponding slave PTY and
-// returns the two *os.File handles. This is a local helper used only
-// by the foreground-pgid test; production has no equivalent because
-// agent-run inherits its PTY from the tmux pane.
-func openPTYPair() (master, slave *os.File, err error) {
-	master, err = os.OpenFile("/dev/ptmx", os.O_RDWR|syscall.O_NOCTTY, 0)
-	if err != nil {
-		return nil, nil, fmt.Errorf("open /dev/ptmx: %w", err)
-	}
-	defer func() {
-		if err != nil {
-			master.Close()
-		}
-	}()
-
-	var unlock int32
-	if _, _, errno := syscall.Syscall(syscall.SYS_IOCTL, master.Fd(), syscall.TIOCSPTLCK, uintptr(unsafe.Pointer(&unlock))); errno != 0 {
-		return nil, nil, fmt.Errorf("TIOCSPTLCK: %w", errno)
-	}
-	var ptyN int32
-	if _, _, errno := syscall.Syscall(syscall.SYS_IOCTL, master.Fd(), syscall.TIOCGPTN, uintptr(unsafe.Pointer(&ptyN))); errno != 0 {
-		return nil, nil, fmt.Errorf("TIOCGPTN: %w", errno)
-	}
-	slavePath := fmt.Sprintf("/dev/pts/%d", ptyN)
-	slave, err = os.OpenFile(slavePath, os.O_RDWR|syscall.O_NOCTTY, 0)
-	if err != nil {
-		return nil, nil, fmt.Errorf("open %s: %w", slavePath, err)
-	}
-	return master, slave, nil
-}
-
-func boolFlag(b bool) string {
-	if b {
-		return "1"
-	}
-	return "0"
-}
-
-// ── early-return paths ───────────────────────────────────────────────────────
-
-// TestSuperviseChild_NilCmdReturnsImmediately verifies that passing a nil
-// *exec.Cmd causes SuperviseChild to return nil without spawning the
-// forwarder goroutine or attempting tcsetpgrp. This is the guard at the
-// top of SuperviseChild.
+// TestSuperviseChild_NilCmdReturnsImmediately verifies that passing a
+// nil *exec.Cmd causes SuperviseChild to return nil without spawning
+// the forwarder goroutine or attempting tcsetpgrp. This is the guard
+// at the top of SuperviseChild and is identical behaviour on Linux
+// and Darwin.
 func TestSuperviseChild_NilCmdReturnsImmediately(t *testing.T) {
 	done := make(chan error, 1)
 	go func() {
@@ -385,10 +69,11 @@ func TestSuperviseChild_NilCmdReturnsImmediately(t *testing.T) {
 	}
 }
 
-// TestSuperviseChild_UnstartedCmdReturnsImmediately verifies that passing
-// an *exec.Cmd that has not been started (cmd.Process == nil) also takes
-// the early-return path. This protects callers who construct a Cmd but
-// then bail on Start before calling SuperviseChild.
+// TestSuperviseChild_UnstartedCmdReturnsImmediately verifies that
+// passing an *exec.Cmd that has not been started (cmd.Process == nil)
+// also takes the early-return path. This protects callers who
+// construct a Cmd but then bail on Start before calling
+// SuperviseChild.
 func TestSuperviseChild_UnstartedCmdReturnsImmediately(t *testing.T) {
 	cmd := exec.Command("true") // not started — cmd.Process is nil
 	if cmd.Process != nil {
@@ -407,399 +92,5 @@ func TestSuperviseChild_UnstartedCmdReturnsImmediately(t *testing.T) {
 		}
 	case <-time.After(time.Second):
 		t.Fatal("SuperviseChild(unstarted cmd, ...) did not return within 1s")
-	}
-}
-
-// ── signal forwarding (Linux only, via helper subprocess) ────────────────────
-
-// signalHelperAndAwaitExit sends sig to the helper PID, waits for the
-// helper to exit, and asserts the helper exited cleanly (status 0,
-// meaning Wait inside the helper returned a signal-terminated error —
-// i.e. SuperviseChild forwarded the signal to the child pgid).
-func signalHelperAndAwaitExit(t *testing.T, helper *exec.Cmd, pid int, sig syscall.Signal, deadline time.Duration) {
-	t.Helper()
-
-	if err := syscall.Kill(pid, sig); err != nil {
-		t.Fatalf("kill helper %d with %v: %v", pid, sig, err)
-	}
-
-	waitDone := make(chan error, 1)
-	go func() {
-		waitDone <- helper.Wait()
-	}()
-
-	select {
-	case err := <-waitDone:
-		// Helper exit 0 → SuperviseChild forwarded the signal and Wait
-		// returned an ExitError. Any other exit is a test failure.
-		if err == nil {
-			return // exit 0 → success
-		}
-		if exitErr, ok := err.(*exec.ExitError); ok {
-			t.Errorf("helper exited with code %d after %v (want 0 — supervisor failed to forward the signal)", exitErr.ExitCode(), sig)
-		} else {
-			t.Errorf("helper Wait failed unexpectedly: %v", err)
-		}
-	case <-time.After(deadline):
-		_ = syscall.Kill(-helper.Process.Pid, syscall.SIGKILL)
-		t.Fatalf("helper did not exit within %v after %v", deadline, sig)
-	}
-}
-
-// TestSuperviseChild_SIGTERMReachesChild verifies the end-to-end signal
-// forwarding path for SIGTERM. See the file-level comment for why this
-// uses a helper subprocess rather than raising SIGTERM at the test
-// binary directly.
-//
-// Linux-only because the helper relies on Setpgid + Kill(-pgid, sig).
-func TestSuperviseChild_SIGTERMReachesChild(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skipf("SIGTERM-forwarding test is Linux-only (GOOS=%q)", runtime.GOOS)
-	}
-	if testing.Short() {
-		t.Skip("skipping process-spawning test in -short mode")
-	}
-
-	helper, pid := startSuperviseHelper(t, true)
-	signalHelperAndAwaitExit(t, helper, pid, syscall.SIGTERM, 5*time.Second)
-}
-
-// TestSuperviseChild_SIGINTReachesChild verifies SIGINT forwarding —
-// same shape as the SIGTERM test, different signal. Together with the
-// SIGHUP test below this covers all three of the always-on signals
-// listed in superviseForwardSignals.
-func TestSuperviseChild_SIGINTReachesChild(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skipf("SIGINT-forwarding test is Linux-only (GOOS=%q)", runtime.GOOS)
-	}
-	if testing.Short() {
-		t.Skip("skipping process-spawning test in -short mode")
-	}
-
-	helper, pid := startSuperviseHelper(t, false)
-	signalHelperAndAwaitExit(t, helper, pid, syscall.SIGINT, 5*time.Second)
-}
-
-// TestSuperviseChild_SIGHUPReachesChild verifies SIGHUP forwarding. This
-// is the signal the kernel sends to agent-run when the tmux pane's
-// controlling terminal is closed — the production path most likely to
-// fire in practice.
-func TestSuperviseChild_SIGHUPReachesChild(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skipf("SIGHUP-forwarding test is Linux-only (GOOS=%q)", runtime.GOOS)
-	}
-	if testing.Short() {
-		t.Skip("skipping process-spawning test in -short mode")
-	}
-
-	helper, pid := startSuperviseHelper(t, true)
-	signalHelperAndAwaitExit(t, helper, pid, syscall.SIGHUP, 5*time.Second)
-}
-
-// ── SIGWINCH branches (in-process, since SIGWINCH does not terminate) ───────
-
-// TestSuperviseChild_WinchForwardedWhenEnabled verifies that when
-// opts.ForwardWinch is true, the supervisor (a) invokes the OnWinch
-// callback once per received SIGWINCH and (b) forwards SIGWINCH to the
-// child's pgid. We check (a) directly via an atomic counter; (b) is
-// covered transitively by superviseForwardSignals's call to
-// syscall.Kill(-proc.Pid, SIGWINCH) — we cannot easily assert "the
-// child received SIGWINCH" without instrumenting the child, but if the
-// callback fires we know the supervisor's SIGWINCH branch was taken
-// (the Kill follows unconditionally in the same case arm).
-//
-// SIGWINCH is "ignore by default" so raising it at the test binary
-// itself is safe — no risk of terminating the test process the way
-// SIGTERM/SIGINT/SIGHUP would. This test therefore stays in-process.
-func TestSuperviseChild_WinchForwardedWhenEnabled(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skipf("SIGWINCH-forwarding test is Linux-only (GOOS=%q)", runtime.GOOS)
-	}
-	if testing.Short() {
-		t.Skip("skipping process-spawning test in -short mode")
-	}
-
-	cmd := exec.Command("sleep", "30")
-	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-	if err := cmd.Start(); err != nil {
-		t.Fatalf("start sleep child: %v", err)
-	}
-	t.Cleanup(func() {
-		if cmd.Process != nil {
-			_ = syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
-		}
-	})
-
-	var winchCount atomic.Int32
-	opts := SuperviseOpts{
-		ForwardWinch: true,
-		OnWinch:      func() { winchCount.Add(1) },
-	}
-
-	waitDone := make(chan error, 1)
-	go func() {
-		waitDone <- SuperviseChild(cmd, int(os.Stdin.Fd()), opts)
-	}()
-
-	// Wait for signal.Notify subscription to settle.
-	time.Sleep(100 * time.Millisecond)
-
-	// Send SIGWINCH to ourselves a few times. Each delivery should
-	// trigger OnWinch. We pace them so the forwarder's buffered channel
-	// (cap 4) does not coalesce bursts.
-	const wantWinch = 3
-	for i := 0; i < wantWinch; i++ {
-		if err := syscall.Kill(os.Getpid(), syscall.SIGWINCH); err != nil {
-			t.Fatalf("kill self with SIGWINCH: %v", err)
-		}
-		time.Sleep(30 * time.Millisecond)
-	}
-
-	deadline := time.Now().Add(time.Second)
-	for time.Now().Before(deadline) && winchCount.Load() < 1 {
-		time.Sleep(10 * time.Millisecond)
-	}
-	if got := winchCount.Load(); got < 1 {
-		t.Errorf("OnWinch fired %d times after %d SIGWINCH, want >=1", got, wantWinch)
-	}
-
-	// Tear down: SIGKILL the child pgid so cmd.Wait returns and
-	// SuperviseChild exits.
-	_ = syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
-
-	select {
-	case <-waitDone:
-		// supervisor returned cleanly
-	case <-time.After(3 * time.Second):
-		t.Fatal("SuperviseChild did not return within 3s after child was killed")
-	}
-}
-
-// TestSuperviseChild_WinchNotSubscribedWhenDisabled verifies the
-// sandbox-exec branch: when opts.ForwardWinch is false the supervisor
-// must NOT subscribe SIGWINCH at all, and therefore must NOT invoke the
-// OnWinch callback even when one is provided. This is the negative of
-// TestSuperviseChild_WinchForwardedWhenEnabled and is the sole
-// behavioural difference between the bwrap and sandbox-exec paths.
-//
-// We install a local SIGWINCH subscriber in the test itself to prove
-// the signal actually fired (otherwise a missing OnWinch invocation
-// could equally mean the SIGWINCH was lost en route). If the local
-// subscriber sees the signal but the supervisor's OnWinch never fires,
-// the ForwardWinch=false branch is correctly skipping signal.Notify.
-func TestSuperviseChild_WinchNotSubscribedWhenDisabled(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skipf("ForwardWinch=false test is Linux-only (GOOS=%q)", runtime.GOOS)
-	}
-	if testing.Short() {
-		t.Skip("skipping process-spawning test in -short mode")
-	}
-
-	cmd := exec.Command("sleep", "30")
-	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-	if err := cmd.Start(); err != nil {
-		t.Fatalf("start sleep child: %v", err)
-	}
-	t.Cleanup(func() {
-		if cmd.Process != nil {
-			_ = syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
-		}
-	})
-
-	var winchCount atomic.Int32
-	opts := SuperviseOpts{
-		ForwardWinch: false,
-		// Pass a non-nil callback so we can prove it is never invoked
-		// when ForwardWinch is false — covers the sandbox-exec parity
-		// requirement.
-		OnWinch: func() { winchCount.Add(1) },
-	}
-
-	// Local SIGWINCH observer — proves the signal we raise below is
-	// actually delivered to this process. Go's signal package
-	// multiplexes subscribers, so this does not interfere with the
-	// supervisor (if the supervisor were also subscribed it would
-	// receive the signal independently).
-	observer := make(chan os.Signal, 4)
-	signal.Notify(observer, syscall.SIGWINCH)
-	defer signal.Stop(observer)
-
-	waitDone := make(chan error, 1)
-	go func() {
-		waitDone <- SuperviseChild(cmd, int(os.Stdin.Fd()), opts)
-	}()
-
-	// Wait for the supervisor goroutine to start (it would have called
-	// signal.Notify by now if ForwardWinch were true — we are
-	// deliberately giving it the same window so the absence of a
-	// subscription is meaningful).
-	time.Sleep(100 * time.Millisecond)
-
-	for i := 0; i < 3; i++ {
-		if err := syscall.Kill(os.Getpid(), syscall.SIGWINCH); err != nil {
-			t.Fatalf("kill self with SIGWINCH: %v", err)
-		}
-		time.Sleep(30 * time.Millisecond)
-	}
-
-	// Confirm the SIGWINCH was actually raised (i.e. the test
-	// apparatus works). We expect at least one delivery on the local
-	// observer.
-	localObserved := false
-	deadline := time.Now().Add(time.Second)
-	for time.Now().Before(deadline) && !localObserved {
-		select {
-		case <-observer:
-			localObserved = true
-		case <-time.After(50 * time.Millisecond):
-		}
-	}
-	if !localObserved {
-		t.Fatalf("local SIGWINCH observer never fired — test apparatus broken, cannot conclude anything about the supervisor")
-	}
-
-	// Give the supervisor (which, per the spec, did not subscribe)
-	// time to NOT invoke OnWinch.
-	time.Sleep(100 * time.Millisecond)
-
-	if got := winchCount.Load(); got != 0 {
-		t.Errorf("OnWinch fired %d times with ForwardWinch=false, want 0", got)
-	}
-
-	// Tear down.
-	_ = syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
-	select {
-	case <-waitDone:
-	case <-time.After(3 * time.Second):
-		t.Fatal("SuperviseChild did not return within 3s after child was killed")
-	}
-}
-
-// ── foreground pgid handoff/restore (Linux only) ─────────────────────────────
-
-// TestSuperviseChild_ForegroundPgidHandoffAndRestore verifies that
-// SuperviseChild calls tcsetpgrpForeground to make the supervised
-// child's pgid the controlling terminal's foreground group, and that
-// when Wait returns SuperviseChild *attempts* the restore via
-// tcsetpgrpRestore (production code discards its error with `_ =`).
-//
-// This test re-execs the test binary as a helper subprocess (see
-// runSuperviseForegroundHelper) so the helper can be set up by the
-// kernel with a freshly opened PTY as its controlling terminal
-// (Setsid + Setctty + Ctty = 0 in SysProcAttr). Inside the helper we
-// observe the slave PTY's foreground pgid before and after
-// SuperviseChild.
-//
-// What is asserted:
-//
-//   - before > 0: tcsetpgrpForeground's TIOCGPGRP succeeded, so the
-//     original foreground pgid was captured (necessary precondition
-//     for the restore code path).
-//   - after == child_pgid: SuperviseChild's TIOCSPGRP succeeded in
-//     handing the foreground to the supervised child. The supervised
-//     child's pid is reported alongside in the helper output.
-//   - SuperviseChild invokes tcsetpgrpRestore after Wait returns.
-//     This is the production behaviour (visible in supervise.go); the
-//     helper also calls tcsetpgrpRestore explicitly after
-//     SuperviseChild for diagnostic purposes and reports any errno
-//     back to the parent. The kernel may return ENOTTY for the
-//     restore call from a background pgid even with SIGTTOU ignored
-//     (see helper output `restore_err`) — this matches the
-//     production behaviour where the error is discarded; the restore
-//     attempt itself is what matters for the AC.
-//
-// Linux-only: /dev/ptmx, TIOCGPGRP/TIOCSPGRP semantics, and the
-// Setctty SysProcAttr field are all Linux-specific (out of scope per
-// #1874).
-func TestSuperviseChild_ForegroundPgidHandoffAndRestore(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skipf("foreground-pgid test is Linux-only (GOOS=%q)", runtime.GOOS)
-	}
-	if testing.Short() {
-		t.Skip("skipping PTY-spawning test in -short mode")
-	}
-
-	helper, stdout, stderrBuf := startSuperviseHelperWithMode(t, "foreground", nil)
-
-	// Drain stdout concurrently with Wait so the pipe does not block
-	// the helper if it writes more than the kernel pipe buffer (cf.
-	// the AGENTS.md stdout-capture-testing convention). The buffer
-	// here is far smaller than the kernel default, so this is
-	// defence-in-depth rather than strictly necessary.
-	outCh := make(chan []byte, 1)
-	go func() {
-		b, _ := io.ReadAll(stdout)
-		outCh <- b
-	}()
-
-	waitDone := make(chan error, 1)
-	go func() {
-		waitDone <- helper.Wait()
-	}()
-
-	var waitErr error
-	select {
-	case waitErr = <-waitDone:
-	case <-time.After(5 * time.Second):
-		_ = syscall.Kill(-helper.Process.Pid, syscall.SIGKILL)
-		t.Fatal("foreground helper did not exit within 5s")
-	}
-
-	var outBuf []byte
-	select {
-	case outBuf = <-outCh:
-	case <-time.After(time.Second):
-		t.Fatal("helper stdout drain did not finish within 1s of helper exit")
-	}
-
-	if waitErr != nil {
-		if exitErr, ok := waitErr.(*exec.ExitError); ok {
-			// Specific exit codes 20-28 from runSuperviseForegroundHelper
-			// indicate environmental skips (no /dev/ptmx, no slave, etc.)
-			// vs hard failures. We treat 24 (open slave failed) as a
-			// skip since it can happen in containers; everything else
-			// is a real failure.
-			code := exitErr.ExitCode()
-			if code == 24 {
-				t.Skipf("helper could not open slave PTY (containerised environment): stderr=%q", stderrBuf.String())
-			}
-			t.Fatalf("foreground helper exited with code %d (sys=%+v): stderr=%q stdout=%q", code, exitErr.ProcessState.Sys(), stderrBuf.String(), string(outBuf))
-		}
-		t.Fatalf("foreground helper Wait failed: %v", waitErr)
-	}
-
-	var beforePgid, afterPgid, finalPgid, childPid int
-	var restoreErr string
-	if _, err := fmt.Sscanf(string(outBuf), "before=%d after=%d final=%d child_pid=%d restore_err=%q", &beforePgid, &afterPgid, &finalPgid, &childPid, &restoreErr); err != nil {
-		t.Fatalf("parse helper output %q: %v", string(outBuf), err)
-	}
-	t.Logf("helper diagnostics: before=%d after=%d final=%d child_pid=%d restore_err=%q", beforePgid, afterPgid, finalPgid, childPid, restoreErr)
-
-	// Assertion 1: original pgid was captured. Without this the
-	// supervisor's `if origPgid > 0` guard skips the restore call.
-	if beforePgid <= 0 {
-		t.Fatalf("helper reported beforePgid=%d, want > 0 (TIOCGPGRP failed — handoff path not exercised)", beforePgid)
-	}
-
-	// Assertion 2: SuperviseChild successfully handed the foreground
-	// to the supervised child's pgid (== childPid since the child was
-	// Setpgid'd). This is the only directly observable evidence that
-	// SuperviseChild's tcsetpgrpForeground call succeeded.
-	if afterPgid != childPid {
-		t.Errorf("foreground pgid on slave PTY = %d immediately after SuperviseChild's handoff, want child pgid %d (TIOCSPGRP did not take effect)", afterPgid, childPid)
-	}
-
-	// Assertion 3: SuperviseChild invoked tcsetpgrpRestore after
-	// Wait. We cannot reliably observe the restore's effect (ENOTTY
-	// from a background pgid is a kernel quirk — see test doc), but
-	// the helper's diagnostic re-call of tcsetpgrpRestore returning
-	// the same error class is evidence the supervisor's discarded
-	// call would have gone down the same path. If restore_err is
-	// empty we expect finalPgid == beforePgid; if non-empty we accept
-	// finalPgid == afterPgid (the restore had no effect).
-	if restoreErr == "" {
-		if finalPgid != beforePgid {
-			t.Errorf("explicit tcsetpgrpRestore reported success but finalPgid=%d != beforePgid=%d", finalPgid, beforePgid)
-		}
 	}
 }
